requirements.txt
-e .
aws-cdk.aws_apigateway
aws_cdk.aws_dynamodb
aws-cdk.aws_lambda
aws-cdk.aws-s3
aws-cdk.aws-iam
jsii